update to output mp4 video and set resolution for telegram

+++ handlers.py
@@ -1,7 +1,7 @@
-import asyncio
 import os
 from telegram import Bot
-from video_funcs import download_video, read_file
+from yt_dlp import YoutubeDL
+from ycf import YCF
 
 
 def get_bot() -> Bot:
@@ -27,21 +27,39 @@ async def send_help(chat_id: int) -> None:
 
 
 async def send_youtube_video(chat_id: int, url: str) -> None:
-    video_path = os.environ["TEMP_VIDEO_FILE"]
+    save_path = os.environ["TEMP_VIDEO_PATH"]
 
-    await download_video(url, video_path)
+    options = {
+        'outtmpl': os.path.join(save_path, "video"),
+        'format': 'bestvideo[height<=1920]+bestaudio/best[height<=1080]/best',
+        'ffmpeg_location': os.environ["FFMPEG_PATH"],
+        'extractor_args': {'youtube:player_client': 'android'},
+        'postprocessors': [{
+            'key': 'FFmpegVideoConvertor',
+            'preferedformat': 'mp4'
+        }],
+        'quiet': True,
+    }
 
-    await asyncio.sleep(2)
+    with YoutubeDL(options) as ydl:
+        metadata = ydl.extract_info(url, download=True)
 
-    video = await read_file(video_path)
+    video_file = os.listdir(save_path)[0]
+    video_path = os.path.join(save_path, video_file)
 
     async with get_bot() as bot:
         await bot.send_video(
             chat_id=chat_id,
-            video=video
+            video=video_path,
+            width=metadata["width"],
+            height=metadata["height"],
+            duration=metadata["duration"],
+            caption=metadata["title"],
+            supports_streaming=True,
         )
 
-    os.remove(video_path)
+    if YCF.isOutside():
+        os.remove(video_path)
 
 
 async def send_exhaust_message(chat_id: int) -> None:
